pyproject.toml
[tool.poetry]
name = "script-py"
version = "0.1.0"
description = ""
authors = ["dengxiaojun <dxj1718874198@gmail.com>"]

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"

[tool.poetry.dependencies]
python = "^3.6"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
script-py = "script_py.Application:Application.run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
